Rename BasePage as _BasePage and introduce simple StatelessWidget BasePage.
