+++ lib/main.dart
@@ -8,11 +8,11 @@ import 'package:kar_kam/kar_kam.dart';
 
 //  App start point.
 void main() {
-  // //  Define an instance of GetIt and register AppSettings.
-  // GetIt.instance.registerSingleton<AppSettings>(
-  //   AppSettings(),
-  //   signalsReady: true,
-  // );
+  //  Define an instance of GetIt and register AppSettings.
+  GetIt.instance.registerSingleton<AppSettings>(
+    AppSettings(),
+    signalsReady: true,
+  );
 
   //  Run the app.
   runApp(const KarKam());
